Refactor tests: move batchSequencing test to TestPostgres15 only and enable previously disabled JooQ tests (#5)
