+++ transactionoutbox-testing/src/main/java/com/gruelbox/transactionoutbox/testing/AbstractAcceptanceTest.java
@@ -33,8 +33,8 @@ public abstract class AbstractAcceptanceTest extends BaseTest {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(AbstractAcceptanceTest.class);
 
-  private ExecutorService unreliablePool;
-  private ExecutorService singleThreadPool;
+  protected ExecutorService unreliablePool;
+  protected ExecutorService singleThreadPool;
 
   private static final Random random = new Random();
 
@@ -111,112 +111,6 @@ public abstract class AbstractAcceptanceTest extends BaseTest {
             .listener(new LatchListener(latch))
             .initializeImmediately(false)
             .flushBatchSize(4)
-            .build();
-
-    outbox.initialize();
-    clearOutbox();
-
-    withRunningFlusher(
-        outbox,
-        () -> {
-          transactionManager.inTransaction(
-              () -> {
-                for (int i = 1; i <= countPerTopic; i++) {
-                  for (int j = 1; j <= topicCount; j++) {
-                    outbox
-                        .with()
-                        .ordered("topic" + j)
-                        .schedule(InterfaceProcessor.class)
-                        .process(i, "topic" + j);
-                  }
-                }
-              });
-          assertTrue(latch.await(30, SECONDS));
-        });
-
-    var output = new HashMap<String, ArrayList<Integer>>();
-    transactionManager.inTransaction(
-        tx -> {
-          //noinspection resource
-          try (var stmt = tx.connection().createStatement();
-              var rs = stmt.executeQuery("SELECT topic, foo FROM TEST_TABLE ORDER BY ix")) {
-            while (rs.next()) {
-              ArrayList<Integer> values =
-                  output.computeIfAbsent(rs.getString(1), k -> new ArrayList<>());
-              values.add(rs.getInt(2));
-            }
-          } catch (SQLException e) {
-            throw new RuntimeException(e);
-          }
-        });
-
-    var indexes = IntStream.range(1, countPerTopic + 1).boxed().collect(toList());
-    var expected =
-        IntStream.range(1, topicCount + 1)
-            .mapToObj(i -> "topic" + i)
-            .collect(toMap(it -> it, it -> indexes));
-    assertEquals(expected, output);
-  }
-
-  @Test
-  final void batchSequencing() throws Exception {
-    int countPerTopic = 20;
-    int topicCount = 5;
-
-    AtomicInteger insertIndex = new AtomicInteger();
-    CountDownLatch latch = new CountDownLatch(countPerTopic * topicCount);
-    ThreadLocalContextTransactionManager transactionManager =
-        (ThreadLocalContextTransactionManager) txManager();
-
-    transactionManager.inTransaction(
-        tx -> {
-          //noinspection resource
-          try (var stmt = tx.connection().createStatement()) {
-            stmt.execute("DROP TABLE TEST_TABLE");
-          } catch (SQLException e) {
-            // ignore
-          }
-        });
-
-    transactionManager.inTransaction(
-        tx -> {
-          //noinspection resource
-          try (var stmt = tx.connection().createStatement()) {
-            stmt.execute(createTestTable());
-          } catch (SQLException e) {
-            throw new RuntimeException(e);
-          }
-        });
-
-    TransactionOutbox outbox =
-        TransactionOutbox.builder()
-            .transactionManager(transactionManager)
-            .submitter(Submitter.withExecutor(unreliablePool))
-            .attemptFrequency(Duration.ofMillis(500))
-            .instantiator(
-                new RandomFailingInstantiator(
-                    (foo, bar) -> {
-                      transactionManager.requireTransaction(
-                          tx -> {
-                            //noinspection resource
-                            try (var stmt =
-                                tx.connection()
-                                    .prepareStatement(
-                                        "INSERT INTO TEST_TABLE (topic, ix, foo) VALUES(?, ?, ?)")) {
-                              stmt.setString(1, bar);
-                              stmt.setInt(2, insertIndex.incrementAndGet());
-                              stmt.setInt(3, foo);
-                              stmt.executeUpdate();
-                            } catch (SQLException e) {
-                              throw new RuntimeException(e);
-                            }
-                          });
-                    }))
-            .persistor(persistor())
-            .listener(new LatchListener(latch))
-            .initializeImmediately(false)
-            .flushBatchSize(4)
-            .useOrderedBatchProcessing(true)
             .build();
 
     outbox.initialize();
@@ -887,15 +781,15 @@ public abstract class AbstractAcceptanceTest extends BaseTest {
     }
   }
 
-  private static class RandomFailingInstantiator implements Instantiator {
+  public static class RandomFailingInstantiator implements Instantiator {
 
     private final InterfaceProcessor interfaceProcessor;
 
-    RandomFailingInstantiator() {
+    public RandomFailingInstantiator() {
       this.interfaceProcessor = (foo, bar) -> {};
     }
 
-    RandomFailingInstantiator(InterfaceProcessor interfaceProcessor) {
+    public RandomFailingInstantiator(InterfaceProcessor interfaceProcessor) {
       this.interfaceProcessor = interfaceProcessor;
     }
